+++ src/renderer/features/wallet-details/ui/components/Collapsible.tsx
