+++ src/renderer/features/wallet-details/ui/components/ConsensusAccountsList.tsx
@@ -8,10 +8,10 @@ import { useDeferredList } from '@/shared/lib/hooks';
 import { RelayChains } from '@/shared/lib/utils';
 import { type AccountId } from '@/shared/polkadotjs-schemas';
 import { FootnoteText } from '@/shared/ui';
+import { Account } from '@/shared/ui-entities';
 import { ScrollArea } from '@/shared/ui-kit';
 import { ChainTitle } from '@/entities/chain';
 import { networkUtils } from '@/entities/network';
-import { Account } from '../Account/Account';
 
 import { Collapsible } from './Collapsible';
