+++ src/renderer/features/wallet-details/ui/wallets/SimpleWalletDetails.tsx
@@ -7,7 +7,7 @@ import { useI18n } from '@/shared/i18n';
 import { useModalClose, useToggle } from '@/shared/lib/hooks';
 import { isEthereumAccountId, toAddress } from '@/shared/lib/utils';
 import { HeadlineText, IconButton, Separator } from '@/shared/ui';
-import { ChainAccountsList, ConsensusAccountsList, WalletAccountIcon } from '@/shared/ui-entities';
+import { ChainAccountsList, WalletAccountIcon } from '@/shared/ui-entities';
 import { Box, Modal, ScrollArea, Tabs } from '@/shared/ui-kit';
 import { type AnyAccount, accountService, accounts } from '@/domains/network';
 import { networkModel, networkUtils } from '@/entities/network';
@@ -18,7 +18,7 @@ import { ForgetWalletConfirm } from '@/features/wallets/ForgetWallet';
 import { RenameWallet } from '@/features/wallets/RenameWallet';
 import { walletDetailsModel } from '../../model/wallet-details-model';
 import { walletProxiesModel } from '../../model/wallet-proxies-model';
-import { WalletFiatBalance } from '../components';
+import { ConsensusAccountsList, WalletFiatBalance } from '../components';
 import { ProxiesCount } from '../components/ProxiesCount';
 import { ProxiesList } from '../components/ProxiesList';
 import { Action, type WalletAction, WalletActions } from '../components/WalletActions';
