+++ src/renderer/shared/ui-entities/index.ts
@@ -15,7 +15,6 @@ export { TransactionDetails } from './TransactionDetails/TransactionDetails';
 export { RankedAccount } from './RankedAccount/RankedAccount';
 export { WalletOnboardingCard } from './WalletOnboardingCard/WalletOnboardingCard';
 export { ChainAccountsList } from './ChainAccountsList/ChainAccountsList';
-export { ConsensusAccountsList } from './ConsensusAccountsList/ConsensusAccountsList';
 export { ChainIcon } from './ChainIcon/ChainIcon';
 export { ChainSelect } from './ChainSelect/ChainSelect';
 export { CollectiveRank, getRankTitle } from './CollectiveRank/CollectiveRank';
